Updated README - Add Photos if You Would Like
This is the full README file for the Fall 14 Final Submission. It should go into detail about all the tests/building methods/details needed for the final submission. If one of you wants to edit this, please do so as there might be some typos and it may not sound professional or the way the TAs and Liz want it.